Improves flexibility of the API client
API client should not be responsible of the content of the payload but rather act as a proxy
between the requester and the API. As a result, possible changes in the implementation are
reduced if (or rather when) the API spec changes.

Also, using object descruction approach in the method signature improves flexibility
and allows better usage of defaults.